refactoring py files

+++ server/users.py
@@ -1,4 +1,4 @@
-from flask import request, Response
+from flask import request, abort, Response
 import json
 from app import app
 from db import *
@@ -35,6 +35,17 @@ def create_user():
                     mimetype='application/json')
 
 
+@app.route('/api/users/<username>', methods=['GET'])
+def get_user(username):
+    user = User.objects(username=username)
+    l_user = user.to_json()
+    decoded = json.loads(l_user)
+    if decoded == []:
+        abort(404)
+    return Response(json.dumps(decoded, sort_keys=False, indent=4),
+                    mimetype='application/json')
+
+
 @app.route('/api/users/<username>', methods=['PUT'])
 def update_user(username):
     user = User.objects(username=username)
@@ -47,8 +58,7 @@ def update_user(username):
                                                   decoded[0]["password"]),
                 "set__email": request.json.get("email",
                                                decoded[0]["email"]),
-                "set__user_id": request.json.get("user_id",
-                                                 decoded[0]["user_id"])})
+                "set__user_id": decoded[0]["user_id"]})
 
     user = User.objects(username=request.json.get("username",
                                                   decoded[0]["username"]))
@@ -63,16 +73,9 @@ def delete_user(username):
     user = User.objects(username=username)
     l_user = user.to_json()
     decoded = json.loads(l_user)
+    if decoded == []:
+        abort(404)
     user.delete()
-    return Response(json.dumps(decoded, sort_keys=False, indent=4),
-                    mimetype='application/json')
-
-
-@app.route('/api/users/<username>', methods=['GET'])
-def get_user(username):
-    user = User.objects(username=username)
-    l_user = user.to_json()
-    decoded = json.loads(l_user)
     return Response(json.dumps(decoded, sort_keys=False, indent=4),
                     mimetype='application/json')
 
@@ -117,6 +120,8 @@ def get_word(username, wordname):
     decoded = json.loads(l_word)
     words = decoded[0]["words"]
     word = [word for word in words if word["word"] == wordname]
+    if word == []:
+        abort(404)
     return Response(json.dumps(word[0], sort_keys=False, indent=4),
                     mimetype='application/json')
 
@@ -128,6 +133,8 @@ def get_word_id(username, word_id):
     decoded = json.loads(l_word)
     words = decoded[0]["words"]
     word = [word for word in words if word["word_id"] == word_id]
+    if word == []:
+        abort(404)
     return Response(json.dumps(word[0], sort_keys=False, indent=4),
                     mimetype='application/json')
 
@@ -138,14 +145,20 @@ def change_word(username, word_id):
     l_user = user.to_json()
     decoded = json.loads(l_user)
     words = decoded[0]["words"]
-    word = Word(word_id=request.json.get("word_id",
-                                         words[word_id-1]["word_id"]),
-                word=request.json.get("word", words[word_id-1]["word"]),
+    wordid_index = [k for k in range(len(words)) if words[k]["word_id"]
+                    == word_id]
+    if [word_id for word in words if word["word_id"] == word_id] == []:
+        abort(404)
+    wordid = [word_id for word in words if word["word_id"] == word_id][0]
+    word = Word(word_id=wordid,
+                word=request.json.get("word", words[wordid_index[0]]["word"]),
                 translation=request.json.get("translation",
-                                             words[word_id-1]["translation"]),
+                                             words[wordid_index[0]]
+                                             ["translation"]),
                 strength=1)
+
     new_word = word.to_dict()
-    words[word_id-1] = new_word
+    words[wordid_index[0]] = new_word
     user.update(**{"set__words": words})
     return Response(json.dumps(new_word, sort_keys=False, indent=4),
                     mimetype='application/json')
@@ -158,6 +171,8 @@ def delete_word(username, word_id):
     decoded = json.loads(l_user)
     words = decoded[0]["words"]
     word = [word for word in words if word["word_id"] == word_id]
+    if word == []:
+        abort(404)
     words.pop(words.index(word[0]))
     user.update(**{"set__words": words})
     return Response(json.dumps(words, sort_keys=False, indent=4),
@@ -186,5 +201,6 @@ def compare(username, original, want):
     words = json.loads(l_word)[0]["words"]
     translation = [word["translation"] for word in words
                    if word["word"] == original]
-
+    if translation == []:
+        abort(404)
     return rate_words(translation[0], want)
